[ACA-3765] [ADF] move reusable versioning logic
