+++ lib/core/directives/version-compatibility.directive.ts
@@ -17,7 +17,6 @@
 
 import { Directive, Input, ViewContainerRef, TemplateRef } from '@angular/core';
 import { VersionCompatibilityService } from '../services/version-compatibility.service';
-import { VersionModel } from '../models/product-version.model';
 
 @Directive({
     selector: '[adf-acs-version]'
@@ -38,45 +37,10 @@ export class VersionCompatibilityDirective {
     }
 
     private validateAcsVersion(requiredVersion: string) {
-        if (requiredVersion && this.isVersionSupported(requiredVersion)) {
+        if (requiredVersion && this.versionCompatibilityService.isVersionSupported(requiredVersion)) {
             this.viewContainer.createEmbeddedView(this.templateRef);
         } else {
             this.viewContainer.clear();
         }
     }
-
-    private parseVersion(version: string): VersionModel {
-        const major = version.split('.')[0];
-        const minor = version.split('.')[1] || '0';
-        const patch = version.split('.')[2] || '0';
-
-        return {
-            major: major,
-            minor: minor,
-            patch: patch
-        } as VersionModel;
-    }
-
-    private isVersionSupported(requiredVersion: string): boolean {
-        const parsedRequiredVersion = this.parseVersion(requiredVersion);
-        const currentVersion = this.versionCompatibilityService.getAcsVersion();
-
-        let versionSupported = false;
-
-        if (currentVersion) {
-            if (+currentVersion.major > +parsedRequiredVersion.major) {
-                versionSupported = true;
-            } else if (currentVersion.major === parsedRequiredVersion.major &&
-                +currentVersion.minor > +parsedRequiredVersion.minor) {
-                versionSupported = true;
-            } else if (currentVersion.major === parsedRequiredVersion.major &&
-                currentVersion.minor === parsedRequiredVersion.minor &&
-                +currentVersion.patch >= +parsedRequiredVersion.patch) {
-                versionSupported = true;
-            }
-        }
-
-        return versionSupported;
-    }
-
 }
